requirements.txt
streamlit==1.35.0
requests==2.31.0
streamlit-lottie==0.0.5
Pillow==10.3.0
rdkit-pypi==2022.9.5
scikit-learn==1.4.2
pandas==2.2.2
py3Dmol==2.0.1
biopython==1.83
beautifulsoup4==4.12.3
openai==1.25.1
rdkit==2023.9.6
joblib==1.4.2
numpy==1.26.4
matplotlib==3.9.0
